* master:
  took out a style sheet
  added about us view and graphs view, found new nap button stored in "sandbox.php", added buttons for new views to nav bar.
  Breast event writing back to database, minus timestamps.
  Added graph and about routes.
  Added id's to buttons and query to grab timestamps.
  Writes to database, minus the timestamps.
  Converted form from html to blade.
  Changed bottle_ounces column from decimal to integer.
  Added function and route for bottle.
  Links to diaper form with correct baby id!!!
  Added showMenu function to redirect back to the menu after diaper form is submitted.
  Added redirect after form is submitted.
  Replaced button hash with url.
  Modified form notes.
  Added a redirect back to the menu.

+++ app/database/migrations/2014_07_14_204317_create_eating_table.php
@@ -24,7 +24,7 @@ class CreateEatingTable extends Migration {
            	$table->boolean('bottle');
            	$table->timestamp('start_bottle')->nullable();
            	$table->timestamp('stop_bottle')->nullable();
-           	$table->decimal('bottle_ounces', 4,2)->nullable();
+           	$table->integer('bottle_ounces')->nullable();
             $table->text('notes')->nullable();
             $table->timestamps();
         });
